Migrate datastore to use protos natively, reducing serialization.
For example, Athlete.

Rather than an Athlete model having "height" instead an Athlete model
has an Athlete (proto) property.

The Athlete (proto) message will have the height field.

Two changes as a result: Nothing is indexed by default.
We don't have to re-serialize data when we send it to clients.